shader_recompiler: Implement V_MOVRELS_B32
This inst indexes the vgpr's like an array to read and write at a
dynamic index.
Generates a ton of OpSelects to hardcode reading from the VGPR at src
index if the index is equal to 1..max_allocated_vgpr. Then does that
again to choose the dest vgpr.

I dont know if theres a better way without significantly reworking the
IR. I think the best solution would be to generate OpSwitch during spirv
emitting, but it would need a case for each VGPR values to load it, then
a case to store back to each VPGR at emit time. You lose that ability after
ssa_rewrite, so I dont think theres a good solution.